requirements.txt
flask
Werkzeug
Flask-Mail
schedule
python-dotenv
sqlitecloud